Handle LinkoutData lazily
Because we're doing these as a dictionary, the logic gets evaluated and we hit
the database regardless of if the caller requested the fields or not.
The whole hash constructed eagerly upfront.

This is compounded with the fact that we can't batch load the entities
because we don't know what they are until runtime (could be sources,
diseases, whatever). That means we're hitting the database lots of
needless times even if the caller just asks for fieldName for instance.o

This wraps the calculated values in lambdas so they won't evaluate
immediately and then calls them if and only if the actual field is
requested.